ocr-pipeline/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
fire==0.6.0
google-generativeai==0.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==2.2.4
openai==1.70.0
pdf2image==1.17.0
pillow==11.1.0
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
